2nd try AI Art Generator/requirements.txt
Flask==3.1.1
Pillow==10.4.0
numpy==2.3.1 